+++ changelog.md
@@ -12,6 +12,7 @@
 * **[Plugins]** Updated 'expl_table' in Splash Damage script.
 * **[Mission Generation]** Also save kneeboards in txt-format, found under "kneeboards" within Retribution's installation folder after pressing take-off.
 * **[Modding]** Support for SW mod v2.55
+* **[Modding]** Support for Spanish & Australian Naval Assets v3.2.0 by desdemicabina
 
 ## Fixes
 * **[New Game Wizard]** Settings would not persist when going back to a previous page.